Assume a role for Route53 record creation
If the user specifies the top-level `AssumRole` key in their properties,
this change now assumes that role before updating the Route 53 records.

The reason that the role is only assumed for the Route 53 client is that
you probably do not want to create the certificate for another account,
whereas the parent zone might be in a different account, requiring you
to assume a role before being able to update records.